Allow to read from named pipes
The -f switch ensures that only plain files are used to read from. But
there are other types of files that are readable, but not plain files. An example
is a named pipe. These should be treated as normal, and I guess the
intention of the tests are to check if we can read from a file.

With this change we allow reading from special files which are tested to
be readable.